Slight click improvement, commented feedback
I left feedback starting with "NOTE:" so that you can ctrl+f through any new comments.
One "note" I missed was regarding timeSlice: If the timeSlice number is too high, the screen lags. If the number is too low, the screen flickers.
Potential solution is to implement "Double buffering", (whatever that might be..) or use another means of updating (got a vague notion that runnable does something like that).